requirements.txt
aiogram==3.13.1
aioschedule
aiomysql
sqlalchemy[asyncio]
python-dotenv
telethon